requirements.txt
flask
plyara
yara-python
pyzmq
pyparsing
redis
requests
